requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.86.0
supabase>=2.9.0
httpx>=0.27.2
python-multipart==0.0.9
langdetect==1.0.9
phonenumbers==8.13.30
requests==2.31.0
python-dateutil==2.8.2 
pytz==2023.3.post1